c2

+++ src/ui/Client.java
@@ -1,13 +1,16 @@
 package ui;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
 
 import Bean.Audio;
 import Bean.Book;
 import Bean.MediaItem;
+import exceptions.ItemException;
 import service.BookManiaLibrary;
 import service.Library;
+import validator.DataValidator;
 
 
 public class Client {
@@ -21,12 +24,13 @@ public class Client {
 	private static String author;
 	
 
-	public static void main(String[] args) {
+	public static void main(String[] args)  {
 		int ch,choice;
 		boolean flag = false;
 		Scanner sc  = new Scanner(System.in);
 		Book b1 = new Book(null, null, 499.0, null, null);
 		Audio a1 = new Audio(null, null, 999.0, null, null);
+		DataValidator dv = new DataValidator();
 		BookManiaLibrary bm1 = new BookManiaLibrary(b1, a1);
 		do{
 			System.out.println("Menu\n1.Insert Details\n2.Search item by id\n3.Update item\n4.Display Items\n5.Delete Item\n6.Exit\nEnter your choice");
@@ -35,12 +39,17 @@ public class Client {
 			{
 				case 1:	
 					flag = false;
-					id = "#"+(unique++);
+					id = ""+(unique++);
 					System.out.println("id: "+id);
 							System.out.println("Book/Audio: (1/0)??");
 							choice = sc.nextInt();
 							System.out.println("Enter the title");	
 							title = sc.next();
+							while(!dv.validateTitle(title))
+							{
+								System.out.println("Please enter a title having length greater than 3 and having no special characters \nEnter the title");	
+								title = sc.next();
+							}
 							System.out.println("Enter the price");
 							price = sc.nextDouble();
 							System.out.println("Enter the year");
@@ -64,37 +73,89 @@ public class Client {
 				case 2:
 					System.out.println("Enter the item id to search for?");
 					id = sc.next();
-					MediaItem m2 = bm1.searchItemById(id);
-					if(m2!=null)
+					if(dv.validateId(id))
 					{
-						System.out.println("Item Found !!!\n"+m2);
+						try
+						{
+						MediaItem m2 = bm1.searchItemById(id);
+						
+						
+							if(m2!=null)
+							{
+								System.out.println("Item Found !!!\n"+m2);
+							}
+							else
+							{
+								System.out.println("Item not found !!");
+						}
+							
+						}
+						catch(ItemException e)
+						{
+							System.out.println("Exception");
+						}
 					}
 					else
 					{
-						System.out.println("Item not found !!");
+						System.out.println("Enter an integer id of length greater than 0");
 					}
+
 					break;
 				case 3:
 					System.out.println("Enter the ID to be updated");
 					id = sc.next();
-					MediaItem m3 = bm1.searchItemById(id);
-					if(m3!=null)
+					if(dv.validateId(id))
 					{
-						//System.out.println("Item Found !!!\n"+m2);
-						bm1.updateItem(m3,id);
+						try
+						{
+							MediaItem m3 = bm1.searchItemById(id);
+							if(m3!=null)
+							{
+								try
+								{
+									bm1.updateItem(m3,id);
+								}
+								catch (ItemException e) {
+									System.out.println("Exception");
+								}
+							}
+							else
+							{
+								System.out.println("Item not found !!");
+							}
+						}
+						catch(ItemException e)
+						{
+							System.out.println("Exception");
+						}
+						
 					}
 					else
 					{
-						System.out.println("Item not found !!");
+						System.out.println("Enter an integer id of length greater than 0");
 					}
 					break;
 				case 4:
 					System.out.println(bm1.getItems());
 					break;
 				case 5:
-					System.out.println("Enter the item id to be deleted?");
-					id = sc.next();
-					bm1.deleteItem(id);
+					
+						System.out.println("Enter the item id to be deleted?");
+						id = sc.next();
+						if(dv.validateId(id))
+						{
+							try
+							{
+								bm1.deleteItem(id);
+							}
+							catch (ItemException e) {
+								System.out.println("Exception");
+							}
+						}
+						else
+						{
+							System.out.println("Enter an integer id of length greater than 0");
+						}
 					break;
 				case 6:
 					break;
